move KeqSeq to a root namespace
