+++ core/src/main/scala/com/github/gchudnov/swearwolf/KeySeq.scala
@@ -1,0 +1,14 @@
+package com.github.gchudnov.swearwolf
+
+import com.github.gchudnov.swearwolf.util.{Point, Size}
+
+trait KeySeq
+
+case object UnknownKeySeq extends KeySeq
+case object PartialKeySeq extends KeySeq
+
+final case class SizeKeySeq(sz: Size)                                                                                              extends KeySeq
+final case class CharKeySeq(ch: Char, mods: Set[KeyModifier] = Set.empty[KeyModifier])                                             extends KeySeq
+final case class CtrlKeySeq(key: KeyCode, mods: Set[KeyModifier] = Set.empty[KeyModifier])                                         extends KeySeq
+final case class MouseKeySeq(pt: Point, button: MouseButton, action: MouseAction, mods: Set[KeyModifier] = Set.empty[KeyModifier]) extends KeySeq
+final case class UnfamiliarKeySeq(bytes: Seq[Byte])                                                                                extends KeySeq
