+++ core/src/main/scala/com/github/gchudnov/swearwolf/package.scala
@@ -1,0 +1,3 @@
+package com.github.gchudnov
+
+package object swearwolf extends KeySeqSyntax {}
