+++ core/src/main/scala/com/github/gchudnov/swearwolf/term/ArrayScreen.scala
@@ -1,6 +1,6 @@
 package com.github.gchudnov.swearwolf.term
 
-import com.github.gchudnov.swearwolf.Screen
+import com.github.gchudnov.swearwolf.{KeySeq, Screen}
 import com.github.gchudnov.swearwolf.util.EventLoop.KeySeqHandler
 import com.github.gchudnov.swearwolf.util.{Point, Size, TextStyle}
