+++ core/src/main/scala/com/github/gchudnov/swearwolf/term/IOTerm.scala
@@ -1,6 +1,8 @@
 package com.github.gchudnov.swearwolf.term
 
-import java.io.{ InputStream, OutputStream }
+import java.io.{InputStream, OutputStream}
+
+import com.github.gchudnov.swearwolf.KeySeq
 
 import scala.collection.mutable.ListBuffer
 import scala.util.control.Exception.nonFatalCatch
