+++ core/src/main/scala/com/github/gchudnov/swearwolf/term/Term.scala
@@ -2,6 +2,8 @@ package com.github.gchudnov.swearwolf.term
 
 import java.io.BufferedOutputStream
 
+import com.github.gchudnov.swearwolf.KeySeq
+
 trait Term {
   def write(bytes: Array[Byte]): Either[Throwable, Unit]
   def write(seq: EscSeq): Either[Throwable, Unit]
