+++ core/src/main/scala/com/github/gchudnov/swearwolf/term/readers/CtrlReader.scala
@@ -1,7 +1,6 @@
 package com.github.gchudnov.swearwolf.term.readers
 
-import com.github.gchudnov.swearwolf.{KeyCode, KeyModifier}
-import com.github.gchudnov.swearwolf.term._
+import com.github.gchudnov.swearwolf.{CharKeySeq, CtrlKeySeq, KeyCode, KeyModifier, KeySeq, UnknownKeySeq}
 
 /**
  * Reads Ctrl + {KEY} sequences.
