+++ core/src/main/scala/com/github/gchudnov/swearwolf/term/readers/EscReader.scala
@@ -1,7 +1,6 @@
 package com.github.gchudnov.swearwolf.term.readers
 
-import com.github.gchudnov.swearwolf.{KeyCode, KeyModifier}
-import com.github.gchudnov.swearwolf.term._
+import com.github.gchudnov.swearwolf.{CtrlKeySeq, KeyCode, KeyModifier, KeySeq, PartialKeySeq, SizeKeySeq, UnknownKeySeq}
 import com.github.gchudnov.swearwolf.util.Size
 
 import scala.annotation.tailrec
