+++ core/src/main/scala/com/github/gchudnov/swearwolf/term/readers/KeySeqReader.scala
@@ -1,6 +1,6 @@
 package com.github.gchudnov.swearwolf.term.readers
 
-import com.github.gchudnov.swearwolf.term.KeySeq
+import com.github.gchudnov.swearwolf.KeySeq
 
 private[term] trait KeySeqReader {
   def read(data: Seq[Byte]): (KeySeq, Seq[Byte])
