+++ core/src/main/scala/com/github/gchudnov/swearwolf/term/readers/MouseReader.scala
@@ -1,7 +1,6 @@
 package com.github.gchudnov.swearwolf.term.readers
 
-import com.github.gchudnov.swearwolf.{KeyModifier, MouseAction, MouseButton}
-import com.github.gchudnov.swearwolf.term._
+import com.github.gchudnov.swearwolf.{KeyModifier, KeySeq, MouseAction, MouseButton, MouseKeySeq, PartialKeySeq, UnknownKeySeq}
 import com.github.gchudnov.swearwolf.util.Point
 
 import scala.annotation.tailrec
