+++ core/src/test/scala/com/github/gchudnov/swearwolf/term/readers/CharReaderSpec.scala
@@ -1,6 +1,6 @@
 package com.github.gchudnov.swearwolf.term.readers
 
-import com.github.gchudnov.swearwolf.term._
+import com.github.gchudnov.swearwolf.{CharKeySeq, KeySeq, UnknownKeySeq}
 import zio.test.Assertion.equalTo
 import zio.test._
