+++ core/src/test/scala/com/github/gchudnov/swearwolf/term/readers/MouseReaderSpec.scala
@@ -1,7 +1,6 @@
 package com.github.gchudnov.swearwolf.term.readers
 
-import com.github.gchudnov.swearwolf.{KeyModifier, MouseAction, MouseButton}
-import com.github.gchudnov.swearwolf.term._
+import com.github.gchudnov.swearwolf.{KeyModifier, KeySeq, MouseAction, MouseButton, MouseKeySeq}
 import com.github.gchudnov.swearwolf.util.{Bytes, Point}
 import zio.test.Assertion._
 import zio.test._
